+++ tomcat/src/main/java/org/apache/catalina/RequestMapping.java
@@ -1,19 +1,19 @@
 package org.apache.catalina;
 
-import org.apache.coyote.controller.AbstractController;
+import nextstep.jwp.controller.AbstractController;
+import nextstep.jwp.controller.DefaultController;
+import nextstep.jwp.controller.LoginController;
+import nextstep.jwp.controller.MemberRegisterController;
 import org.apache.coyote.controller.Controller;
-import org.apache.coyote.controller.EmptyController;
-import org.apache.coyote.controller.LoginHandler;
-import org.apache.coyote.controller.MemberRegisterHandler;
 import org.apache.coyote.http11.request.HttpRequest;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.HashMap;
+import java.util.Map;
 
 public class RequestMapping {
 
     private static RequestMapping instance;
-    private final List<AbstractController> controllers = new ArrayList<>();
+    private final Map<String, Controller> controllers = new HashMap();
 
     public static RequestMapping getInstance() {
         if (instance == null) {
@@ -23,18 +23,14 @@ public class RequestMapping {
     }
 
     private RequestMapping() {
-        final AbstractController loginHandler = new LoginHandler("/login");
-        final AbstractController memberRegisterHandler = new MemberRegisterHandler("/register");
-        controllers.add(loginHandler);
-        controllers.add(memberRegisterHandler);
+        final AbstractController loginHandler = new LoginController();
+        final AbstractController memberRegisterHandler = new MemberRegisterController();
+        controllers.put("/login", loginHandler);
+        controllers.put("/register", memberRegisterHandler);
     }
 
     public Controller getController(HttpRequest request) {
-        for (AbstractController controller : controllers) {
-            if (controller.isMatch(request)) {
-                return controller;
-            }
-        }
-        return new EmptyController();
+        String requestURI = request.getRequestLine().getRequestURI();
+        return controllers.getOrDefault(requestURI, new DefaultController());
     }
 }
